pyproject.toml
[tool.poetry]
name = "threeNineTee"
version = "2.11.1"
description = "A modular, high performance e-commerce storefront built with GraphQL, Django, and ReactJS."
authors = ["Mirumee Software <hello@mirumee.com>"]
license = "BSD-3-Clause"

readme = "README.md"

homepage = "https://getthreeNineTee.com/"
repository = "https://github.com/mirumee/threeNineTee"
documentation = "https://docs.getthreeNineTee.com/en/latest/"

[tool.poetry.dependencies]
python = "~3.8"
babel = "~2.8"
boto3 = "^1.14"
braintree = ">=4.2,<4.5"
celery = { version = ">=4.4.5,<6.0.0", extras = ["redis"] }
dj-database-url = "^0"
dj-email-url = "^1"
django = "^3.0.7"
django-countries = "^6.1"
django-filter = "^2.3"
django-measurement = "^3.0"
django-mptt = "^0"
django-phonenumber-field = ">=4,<6"
django-prices = "^2.1"
django-prices-openexchangerates = "^1.0.1"
django-prices-vatlayer = "^1.1.0"
django-storages = { version = "^1.7.1", extras = [ "google" ] }
django-templated-email = "^2.3.0"
django-versatileimagefield = "^2.0"
draftjs-sanitizer = "^1.0.0"
faker = "^4.1"
freezegun = "^0"
google-cloud-storage = "^1.29.0"
google-i18n-address = "^2.3.5"
google-measurement-protocol = "^1.0"
graphene-django = "^2.11.0"
graphene-federation = "^0.1.0"
html-to-draftjs = "^1.0.1"
markdown = "^3.1.1"
maxminddb = ">=1.5.4,<3.0.0"
maxminddb-geolite2 = "^2018.701"
petl = "1.6.7"
opentracing = "^2.3.0"
phonenumberslite = "^8.12.6"
prices = "^1.0"
psycopg2-binary = "^2.8.3"
purl = "^1.5"
python-magic-bin = {version = "^0.4.14", platform = "win32"}
razorpay = "^1.2"
requests = "^2.22"
sentry-sdk = "^0"
stripe = "^2.47.0"
text-unidecode = "^1.2"
tqdm = "^4.47"
weasyprint = ">=48"
oauthlib = "^3.0"
jaeger-client = "^4.3.0"
openpyxl = "^3.0.3"
django-cache-url = "^3.1.2"
pyjwt = "^1.7.1"
python-json-logger = ">=0.1.11,<2.1.0"
pytimeparse = "^1.1.8"
django-redis = "^4.12.1"
Adyen = "^3.0.0"
google-cloud-pubsub = ">=1.7,<3.0"
uvicorn = {extras = ["standard"], version = "^0.12.2"}
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
black = "19.10b0"
codecov = "^2.1.9"
coverage = "^5.3"
django-debug-toolbar = "^3.1"
django-debug-toolbar-request-history = "^0"
django-graphiql-debug-toolbar = "^0.1.4"
django-extensions = "^3.0.9"
flake8 = "^3.8.4"
isort = "^4.3.21"
pre-commit = "^2.7"
pycodestyle = "^2.5"
pydocstyle = "^5.1"
pylint = "^2.6.0"
pylint-celery = "^0"
pylint-django = "^2.2.0"
pylint-plugin-utils = "^0"
pytest = "^6.1.1"
pytest-celery = "^0.0.0-alpha.1"
pytest-cov = "^2.10.1"
pytest-django = "^3.10.0"
pytest-django-queries = "~1.1"
pytest-mock = "^3.3.1"
pytest-vcr = "^1.0.2"
pytest-xdist = "^1.34"
tox = "^3.19.0"
transifex-client = "^0"
django-stubs = "1.2"
mypy = "0.740"
pywatchman = "^1.4.1"

[tool.black]
target_version = ['py35', 'py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(\.git/
  |\.eggs
  |\.hg
  |__pycache__
  |\.cache
  |\.ipynb_checkpoints
  |\.mypy_cache
  |\.pytest_cache
  |\.tox
  |\.venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |infrastructure
  |templates
  |locale
)/
'''
